Extract ProjectDocumentReference from GroundFirestore
